Database (#5)
* Inicio da implementação da bd (user e betslip)

* Outras tabelas adicionadas, falta testar e queries

* Classes definidas, faltam metodos e queries

* database requirements fix

* Most DB queries

* Update db_main_test.py

* queries

* we grindin

* quase nada xD

* quase nada

* Add currency to betslip

* Update LN

* Bue de cenas

* Muitas cenas. RemoveCurrency

* Merge

* Remove Currency (with TODO)

* Mais funcionalidades db

* Final database changes maybe

* Remove main_test

+++ Server/RASBetLN.py
@@ -1,4 +1,4 @@
-from audioop import tostereo
+
 import RASBetFacade
 from Data import DataBaseAccess
 import datetime, json
@@ -119,7 +119,7 @@ class RASBetLN(RASBetFacade.RASBetFacade):
         return json.dumps(toSend)
 
     def concludeBetSlip(self,userID,amount,currency):
-        amount = int(amount)
+        amount = float(amount)
         success = self.db.withdrawMoney(userID,currency,amount)
 
         toSend = self.createDictWithDefaultInfo(userID)
